requirements.txt
pydantic>=1.9.1
pydantic[dotenv]>=1.9.1
requests>=2.28.0
prettyprint>=0.1.5
discord.py>=2.0.1
coloredlogs>=15.0.1